+++ src/components/main/LineBanner.js
