+++ src/components/main/MainBanner.js
