+++ src/main.js
@@ -1,7 +1,7 @@
-import './components/main/mainBanner.js';
-import './components/main/todayRecommendProducts.js';
-import './components/main/lineBanner.js';
-import './components/main/recentProducts.js';
+import './components/main/MainBanner.js';
+import './components/main/TodayRecommendProducts.js';
+import './components/main/LineBanner.js';
+import './components/main/RecentProducts.js';
 // import './Layout/Header';
 import './Layout/Header/index.js';
 import './Layout/Footer/index.js';
